Moved Mxit header parsing into a separate module.
