+++ lib/mxit-rails.rb
@@ -5,6 +5,7 @@ require "mxit_rails/engine"
 require "mxit_rails/styles"
 require "mxit_rails/controller_extensions"
 require "mxit_rails/validations"
+require "mxit_rails/header"
 require "mxit_rails/page"
 
 module MxitRails
